Fix bugs (#5)
* Fix: category not auto-selected when editing existing item in warehouse

* Fix: part form reused previous ID and updated old record instead of creating a new one

* Add currency to table in class Warehouse

* Fix: Correct decimal formatting for total price in part records

* Refactor: enable proper refresh of warehouse table after adding or editing parts without closed window WarehouseAddEdit

* Fix: Prevent accidental data deletion by adding user confirmation before deleting records in Warehouse table

* Change colour from DimGray to White

* Fix: Prevent saving incomplete vehicle data causing errors in vehicle preview

* Fix: edit text label and some text in generetePDF in program

* Fix: Correct decimal formatting for total price in part records when data transfer beetween 2 tables in class Form_AddRepair. Add Currency to tables

* Change colour from Gray to White

* Update the database to fix incorrect part profit calculation by saving additional data.

* fix incorrect part profit calculation by saving additional data.

* Add currency to 3 table

* Fix visual problems in UC_AddAllData and change CharacterCasing to Upper in RegistrationNumber

* Change logic in UC_AddAllData, Surname is no more required for save if user not input surname

* Fix QuantityItems when adding parts to a car that hasn’t been marked as finished.

* Fix incorrect saving logic payment

+++ Warsztat_2/UserControls/BarMenu/Warehouse/UC_Warehouse.cs
@@ -1,38 +1,35 @@
 ﻿using Warsztat_2.Models;
 
-namespace Warsztat_2._0.UserControls.BarMenu.Warehouse
-{
-    public partial class UC_Warehouse : UserControl
-    {
+namespace Warsztat_2._0.UserControls.BarMenu.Warehouse {
+    public partial class UC_Warehouse :UserControl {
         private WarehouseModel warehouseModel = new WarehouseModel();
-        private readonly WarehouseAddEdit warehouseAddEdit = new();
         private readonly SqlCmd sqlCmd = new();
         public UC_Warehouse()
-        {
+            {
             InitializeComponent();
-        }
+            }
         private async void WarehouseAddButton_Click(object sender, EventArgs e)
-        {
-            warehouseAddEdit.ClearTextBox();
-            warehouseAddEdit.ShowDialog();
-
-            await LoadTable();
-        }
+            {
+            WarehouseAddEdit warehouseAddEditNew = new();
+            warehouseAddEditNew.ClearTextBox();
+            warehouseAddEditNew.RefreshWarehouseTable = async () => await LoadTable();
+            warehouseAddEditNew.ShowDialog();
+            }
 
         private async void UC_Warehouse_Load(object sender, EventArgs e)
-        {
+            {
             await LoadTable();
             await SqlCmd.ReadRecordListBoxAsync(CategorylistBox, "SELECT Typ FROM Kategorie", "Typ");
-        }
+            }
         public async Task LoadTable()
-        {
-            await SqlCmd.LoadData("SELECT ID, Typ, Nazwa, NumerCzęści, Opis, Cena, Ilość, Suma FROM Magazyn", WarehouseView, "Warehouse", "Load table Warehouse From DB");
-            WarehouseView.Columns["ID_Column"].Visible = false;
-        }
-        private void PrepareDataToRead()
-        {
-            warehouseModel = new()
             {
+            await SqlCmd.LoadData("SELECT ID, Typ, Nazwa, NumerCzęści, Opis, Cena, Ilość, ZarobekCzęści, Suma FROM Magazyn", WarehouseView, "Warehouse", "Load table Warehouse From DB");
+            WarehouseView.Columns["ID_Column"].Visible = false;
+            }
+        private void PrepareDataToRead()
+            {
+            warehouseModel = new()
+                {
                 Id = Convert.ToUInt16(WarehouseView.CurrentRow.Cells["ID_Column"].Value.ToString()),
                 Type = WarehouseView.CurrentRow.Cells["Typ_column"].Value.ToString(),
                 PartNumber = WarehouseView.CurrentRow.Cells["NumerCzęści_column"].Value.ToString(),
@@ -40,26 +37,27 @@ namespace Warsztat_2._0.UserControls.BarMenu.Warehouse
                 Description = WarehouseView.CurrentRow.Cells["Opis_column_"].Value.ToString(),
                 Price = Convert.ToDecimal(WarehouseView.CurrentRow.Cells["Cena_column"].Value.ToString()),
                 Quantity = Convert.ToByte(WarehouseView.CurrentRow.Cells["Ilość_column"].Value.ToString()),
+                EarningParts = Convert.ToDecimal(WarehouseView.CurrentRow.Cells["EarningParts_Column"].Value.ToString()),
                 Sum = WarehouseView.CurrentRow.Cells["Suma_Column"].Value.ToString()
-            };
-        }
+                };
+            }
 
         private async void AddCategoryButton_Click(object sender, EventArgs e)
-        {
-            if (!string.IsNullOrWhiteSpace(CategoryTextBox.Text))
             {
+            if(!string.IsNullOrWhiteSpace(CategoryTextBox.Text))
+                {
                 await AddCategory();
+                }
             }
-        }
 
         private async void RemoveCategoryButton_Click(object sender, EventArgs e)
-        {
+            {
             await RemoveCategory();
-        }
+            }
 
         #region Methods of 
         private async Task AddCategory()
-        {
+            {
             var data = new Dictionary<string, object>
             {
                 {"Typ", CategoryTextBox.Text.Trim() }
@@ -69,14 +67,14 @@ namespace Warsztat_2._0.UserControls.BarMenu.Warehouse
 
             CategorylistBox.Items.Add(data["Typ"]);
             CategoryTextBox.Text = string.Empty;
-        }
+            }
         private async Task RemoveCategory()
-        {
-            if (CategorylistBox.SelectedItem == null)
             {
+            if(CategorylistBox.SelectedItem == null)
+                {
                 MessageBox.Show("Wybierz najpierw kategorię dla jej usunięcia :)", "Błąd spowodowany użytkownikiem a nie programistą", MessageBoxButtons.OK, MessageBoxIcon.Warning);
                 return;
-            }
+                }
 
             string categoryToRemove = CategorylistBox.SelectedItem.ToString();
             var whereParams = new Dictionary<string, object>
@@ -86,58 +84,59 @@ namespace Warsztat_2._0.UserControls.BarMenu.Warehouse
 
             await SqlCmd.DeleteRecordAsync("WarsztatDB", "Kategorie", "Typ = @Typ", whereParams);
             CategorylistBox.Items.Remove(categoryToRemove);
-        }
+            }
 
         #endregion
 
         private async void UpdateCategoryList_Click(object sender, EventArgs e)
-        {
+            {
             await SqlCmd.ReadRecordListBoxAsync(CategorylistBox, "SELECT Typ FROM Kategorie", "Typ");
-        }
+            }
 
         private void SearchTextBox_TextChanged(object sender, EventArgs e)
-        {
+            {
             Settings.SearchTextBox(SearchTextBox, WarehouseView);
-        }
+            }
 
         private void CategorylistBox_MouseClick(object sender, MouseEventArgs e)
-        {
-            using BindingSource bindingSource = new()
             {
+            using BindingSource bindingSource = new()
+                {
                 DataSource = WarehouseView.DataSource,
                 Filter = $"{"Typ"} LIKE '%{CategorylistBox.SelectedItem}%'"
-            };
-        }
+                };
+            }
 
         private void CategorylistBox_MouseDoubleClick(object sender, MouseEventArgs e)
-        {
-            using BindingSource bindingSource = new()
             {
+            using BindingSource bindingSource = new()
+                {
                 DataSource = WarehouseView.DataSource
-            };
+                };
             CategorylistBox.SelectedIndices.Clear();
             bindingSource.RemoveFilter();
-        }
+            }
 
         private async void WarehouseView_CellContentClick(object sender, DataGridViewCellEventArgs e)
-        {
-            await SqlCmd.DeleteDataTable(WarehouseView, e, "BtnDelete_", "ID_Column", "Magazyn");
-        }
+            {
+            if(e.ColumnIndex == WarehouseView.Columns["BtnDelete_"].Index &&
+                MessageBox.Show("Na pewno chcesz usunąć te dane?", "Potwierdzenie usunięcia", MessageBoxButtons.YesNo) == DialogResult.Yes)
+                {
+                await SqlCmd.DeleteDataTable(WarehouseView, e, "BtnDelete_", "ID_Column", "Magazyn");
+                }
+            }
 
         private async void WarehouseView_DoubleClick(object sender, EventArgs e)
-        {
-            if (WarehouseView.CurrentRow != null && WarehouseView.CurrentRow.Cells["Typ_column"].Value != DBNull.Value)
             {
+            if(WarehouseView.CurrentRow != null && WarehouseView.CurrentRow.Cells["Typ_column"].Value != DBNull.Value)
+                {
+                WarehouseAddEdit warehouseAddEditNew = new();
                 PrepareDataToRead();
 
-                warehouseAddEdit.SetDataEdit(warehouseModel);
-                warehouseAddEdit.ClearTextBox();
-                warehouseAddEdit.AutocompleteData();
-
-                warehouseAddEdit.ShowDialog();
-
-                await LoadTable();
+                warehouseAddEditNew.SetDataEdit(warehouseModel);
+                warehouseAddEditNew.RefreshWarehouseTable = async () => await LoadTable();
+                warehouseAddEditNew.ShowDialog();
+                }
             }
         }
     }
-}
